TestDB04 working

+++ fr.projetBd/src/main/java/fr/banque/ICompteASeuil.java
@@ -1,12 +1,14 @@
 package fr.banque;
 
+import java.math.BigDecimal;
+
 public interface ICompteASeuil {
 	// En Java 8, on pourrait écrire le corps de cette méthode ici afin
 	// de ne pas dupliquer son code de la classe Compte vers la classe
 	// CompteASeuilRemunere
-	public void retirer(double pMontant) throws BanqueException;
+	public void retirer(BigDecimal pMontant) throws BanqueException;
 
-	public double getSeuil();
+	public BigDecimal getSeuil();
 
-	public void setSeuil(double pSeuil);
+	public void setSeuil(BigDecimal pSeuil);
 }
